#57 Separate examples
