+++ connector/connector_test.go
@@ -94,18 +94,3 @@ func TestNewConnector_mssql(test *testing.T) {
 		assert.FailNow(test, err.Error(), databaseConnector)
 	}
 }
-
-// ----------------------------------------------------------------------------
-// Examples for godoc documentation
-// ----------------------------------------------------------------------------
-
-func ExampleNewConnector_sqlite() {
-	// For more information, visit https://github.com/Senzing/go-databasing/blob/main/connectorpostgresql/connectorpostgresql_test.go
-	ctx := context.TODO()
-	databaseUrl := "sqlite3://na:na@$/tmp/sqlite/G2C.db"
-	databaseConnector, err := NewConnector(ctx, databaseUrl)
-	if err != nil {
-		fmt.Println(err, databaseConnector)
-	}
-	// Output:
-}
