+++ connectormssql/connectormssql_examples_test.go
@@ -1,0 +1,22 @@
+package connectormssql
+
+import (
+	"context"
+	"fmt"
+)
+
+// ----------------------------------------------------------------------------
+// Examples for godoc documentation
+// ----------------------------------------------------------------------------
+
+func ExampleNewConnector() {
+	// For more information, visit https://github.com/Senzing/go-databasing/blob/main/connectormssql/connectormssql_examples_test.go
+	ctx := context.TODO()
+	// See https://github.com/microsoft/go-mssqldb#connection-parameters-and-dsn
+	configuration := "user id=sa;password=Passw0rd;database=master;server=localhost"
+	databaseConnector, err := NewConnector(ctx, configuration)
+	if err != nil {
+		fmt.Println(err, databaseConnector)
+	}
+	// Output:
+}
