+++ connectormssql/connectormssql_test.go
@@ -49,19 +49,3 @@ func TestNewConnector(test *testing.T) {
 		assert.FailNow(test, err.Error(), databaseConnector)
 	}
 }
-
-// ----------------------------------------------------------------------------
-// Examples for godoc documentation
-// ----------------------------------------------------------------------------
-
-func ExampleNewConnector() {
-	// For more information, visit https://github.com/Senzing/go-databasing/blob/main/connectormssql/connectormssql_test.go
-	ctx := context.TODO()
-	// See https://github.com/microsoft/go-mssqldb#connection-parameters-and-dsn
-	configuration := "user id=sa;password=Passw0rd;database=master;server=localhost"
-	databaseConnector, err := NewConnector(ctx, configuration)
-	if err != nil {
-		fmt.Println(err, databaseConnector)
-	}
-	// Output:
-}
