+++ connectormysql/connectormysql_test.go
@@ -57,26 +57,3 @@ func TestNewConnector(test *testing.T) {
 		assert.FailNow(test, err.Error(), databaseConnector)
 	}
 }
-
-// ----------------------------------------------------------------------------
-// Examples for godoc documentation
-// ----------------------------------------------------------------------------
-
-func ExampleNewConnector() {
-	// For more information, visit https://github.com/Senzing/go-databasing/blob/main/connectormysql/connectormysql_test.go
-	ctx := context.TODO()
-	// See https://pkg.go.dev/github.com/go-sql-driver/mysql#Config
-	configuration := &mysql.Config{
-		User:      "root",
-		Passwd:    "root",
-		Net:       "tcp",
-		Addr:      "localhost",
-		Collation: "utf8mb4_general_ci",
-		DBName:    "G2",
-	}
-	databaseConnector, err := NewConnector(ctx, configuration)
-	if err != nil {
-		fmt.Println(err, databaseConnector)
-	}
-	// Output:
-}
