+++ connectorpostgresql/connectorpostgresql_examples_test.go
@@ -1,0 +1,22 @@
+package connectorpostgresql
+
+import (
+	"context"
+	"fmt"
+)
+
+// ----------------------------------------------------------------------------
+// Examples for godoc documentation
+// ----------------------------------------------------------------------------
+
+func ExampleNewConnector() {
+	// For more information, visit https://github.com/Senzing/go-databasing/blob/main/connectorpostgresql/connectorpostgresql_examples_test.go
+	ctx := context.TODO()
+	// See https://pkg.go.dev/github.com/lib/pq#hdr-Connection_String_Parameters
+	configuration := "user=postgres password=postgres dbname=G2 host=localhost sslmode=disable"
+	databaseConnector, err := NewConnector(ctx, configuration)
+	if err != nil {
+		fmt.Println(err, databaseConnector)
+	}
+	// Output:
+}
