+++ connectorpostgresql/connectorpostgresql_test.go
@@ -49,19 +49,3 @@ func TestNewConnector(test *testing.T) {
 		assert.FailNow(test, err.Error(), databaseConnector)
 	}
 }
-
-// ----------------------------------------------------------------------------
-// Examples for godoc documentation
-// ----------------------------------------------------------------------------
-
-func ExampleNewConnector() {
-	// For more information, visit https://github.com/Senzing/go-databasing/blob/main/connectorpostgresql/connectorpostgresql_test.go
-	ctx := context.TODO()
-	// See https://pkg.go.dev/github.com/lib/pq#hdr-Connection_String_Parameters
-	configuration := "user=postgres password=postgres dbname=G2 host=localhost sslmode=disable"
-	databaseConnector, err := NewConnector(ctx, configuration)
-	if err != nil {
-		fmt.Println(err, databaseConnector)
-	}
-	// Output:
-}
