+++ connectorsqlite/connectorsqlite_test.go
@@ -49,18 +49,3 @@ func TestNewConnector(test *testing.T) {
 		assert.FailNow(test, err.Error(), databaseConnector)
 	}
 }
-
-// ----------------------------------------------------------------------------
-// Examples for godoc documentation
-// ----------------------------------------------------------------------------
-
-func ExampleNewConnector() {
-	// For more information, visit https://github.com/Senzing/go-databasing/blob/main/connectorsqlite/connectorsqlite_test.go
-	ctx := context.TODO()
-	configuration := "/tmp/sqlite/G2C.db"
-	databaseConnector, err := NewConnector(ctx, configuration)
-	if err != nil {
-		fmt.Println(err, databaseConnector)
-	}
-	// Output:
-}
